fix(review): harden routing/config/concurrency across 24h audit
近24h提交代码审查发现的缺陷修复：

Critical:
- C1: 注册 ModelsJsonConfigurationSource，dashboard 写 models-config.json 真正生效到路由；
      首启从 appsettings.json Models 段 seed，消除双源 index 合并残留
- C2: 提取 RouterOptionsValidator.ValidateModel，UpsertModel 落盘前校验，
      避免坏配置导致重启 ValidateOnStart 失败

High:
- H1: 流式 size-limit 检查并入内层 try-finally，enumerator 必 dispose，防 socket 泄漏
- H2: fusion cancelledByRace 增加 linkedCts.IsCancellationRequested 判定，
      区分 race-cancel 与 client 真实超时，后者正确计入熔断器
- H3: SQLite 双 store 加 busy_timeout PRAGMA + Default Timeout，防跨 store 写 SQLITE_BUSY
- H4: ConcurrencyRegistry.TrySweep 删除 SemaphoreSlim.Dispose，避免与 GetSemaphore 竞争致 ObjectDisposedException

Medium:
- M1: POST/PUT 数值 clamp，镜像 validator 边界
- M2: /api/models/raw 剥除 ApiKey 明文
- M3: ResetDaily 删除冗余自归档（SnapshotDaily 已归档昨天）
- M4: ModelHealthProbeService 早退改循环内跳过，reload 可重新启用

Low:
- L2: OnFileChanged Reload 移出 _gate 锁
- L3: FileSystemWatcher 增加 Created/Renamed 事件 + NotifyFilter 扩展

其他：
- models-config.json 移除 git 追踪（运行时数据，首启 seed）
- 含 dashboard.js/models.js 输入转义、appsettings TrustProxyHeaders 等预存改动

验证：dotnet build 0 错误，dotnet test 314/314 通过

+++ src/OptiRouter/Endpoints/DashboardHandler.cs
@@ -88,7 +88,7 @@ public static class DashboardHandler
         double qps = recentCount / 60.0;
 
         int totalRequests = recent.Count;
-        int totalTokens = recent.Sum(r => r.PromptTokens + r.CompletionTokens);
+        long totalTokens = recent.Sum(r => (long)r.PromptTokens + r.CompletionTokens);
         double avgLatencyMs = totalRequests > 0 ? recent.Average(r => r.LatencyMs) : 0;
 
         var modelsList = options.Models.Select(m =>
